backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.30
asyncpg==0.29.0
redis==5.0.4
PyJWT==2.8.0
pydantic==2.7.1
python-multipart==0.0.9
httpx==0.27.0

